feat: first try at vgw wasm compat
