+++ src/voice/gateway/mod.rs
@@ -1,0 +1,11 @@
+pub mod backends;
+pub mod events;
+pub mod gateway;
+pub mod handle;
+pub mod heartbeat;
+pub mod message;
+
+pub use backends::*;
+pub use gateway::*;
+pub use handle::*;
+pub use message::*;
